+++ train.py
@@ -86,7 +86,7 @@ def train(data_loader, model):
                 add_tsne_data(tsne_data, features[0])
                 add_tsne_label(tsne_labels, labels.clone()[0])
 
-            loss = criterion(outputs.double(), labels)
+            loss = criterion(outputs.double(), labels.double())
             loss.backward()
 
             optimizer.step()
